+++ conn_test.go
@@ -9,6 +9,7 @@ import (
 	"database/sql/driver"
 	"fmt"
 	"io"
+	"reflect"
 	"strings"
 	"testing"
 	"time"
@@ -18,6 +19,7 @@ import (
 )
 
 func TestParseConnString(t *testing.T) {
+	t.Parallel()
 	wantAt := ConnectionParams{
 		Username:       "cc",
 		Password:       "c@c*1",
@@ -80,6 +82,14 @@ func TestParseConnString(t *testing.T) {
 				WaitTimeout: 30 * time.Second, MaxLifeTime: 1 * time.Hour, SessionTimeout: 5 * time.Minute,
 			},
 		},
+
+		"onInit": {In: "oracle://user:pass@sid/?poolMinSessions=3&poolMaxSessions=9&poolIncrement=3&connectionClass=POOLED&sysoper=1&sysdba=0&poolWaitTimeout=200ms&poolSessionMaxLifetime=4000s&poolSessionTimeout=2000s&onInit=a&onInit=b",
+			Want: ConnectionParams{Username: "user", Password: "pass", SID: "sid",
+				ConnClass: "POOLED", IsSysOper: true,
+				MinSessions: 3, MaxSessions: 9, PoolIncrement: 3,
+				WaitTimeout: 200 * time.Millisecond, MaxLifeTime: 4000 * time.Second, SessionTimeout: 2000 * time.Second,
+				OnInit: []string{"a", "b"},
+			}},
 	} {
 		t.Log(tCase.In)
 		P, err := ParseConnString(tCase.In)
@@ -87,7 +97,7 @@ func TestParseConnString(t *testing.T) {
 			t.Errorf("%s: %v", tName, err)
 			continue
 		}
-		if P != tCase.Want {
+		if !reflect.DeepEqual(P, tCase.Want) {
 			t.Errorf("%s: parse of %q got %#v, wanted %#v\n%s", tName, tCase.In, P, tCase.Want, cmp.Diff(tCase.Want, P))
 			continue
 		}
@@ -97,7 +107,7 @@ func TestParseConnString(t *testing.T) {
 			t.Errorf("%s: parseConnString %v", tName, err)
 			continue
 		}
-		if P != Q {
+		if !reflect.DeepEqual(P, Q) {
 			t.Errorf("%s: params got %+v, wanted %+v\n%s", tName, P, Q, cmp.Diff(P, Q))
 			continue
 		}