+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
+
+## [0.10.0]
+### Added
+- onInit parameter in the connection url (and OnInit in ConnectionParams)
+- export Drv to be able to register new driver wrapping *godror.Drv.
 - ContextWithUserPassw requires a connClass argument, too.
 
 ## [0.9.2]
-### Added
 ### Changed
 - Make Data embed dpiData, not *dpiData